Move logic into pkg for kubeseal
